update: extraindo lógica para rank.rb
